stacktrace: use a 64 bits ID to include potential error
The bpf helper returns the id on 32 bits and the eventual error in the
negatives, so 64 bits can hold both and we can return the information in
the event message.

+++ pkg/api/tracingapi/client_kprobe.go
@@ -49,7 +49,7 @@ type MsgGenericKprobe struct {
 	ActionId     uint64
 	ActionArgId  uint32
 	Tid          uint32 // The recorded TID that triggered the event
-	StackID      uint32
+	StackID      int64
 }
 
 type MsgGenericKprobeArgPath struct {
